+++ internal/services/filestorage/helper.go
@@ -19,8 +19,8 @@ package filestorage
 import (
 	"context"
 
-	coreapi "github.com/unikorn-cloud/core/pkg/openapi"
-	regionapi "github.com/unikorn-cloud/region/pkg/openapi"
+	coreapi "github.com/nscaledev/nscale-sdk-go/common"
+	regionapi "github.com/nscaledev/nscale-sdk-go/region"
 
 	"github.com/nscaledev/terraform-provider-nscale/internal/nscale"
 )