+++ internal/services/instance/instance_flavor_model.go
@@ -19,7 +19,7 @@ package instance
 import (
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	regionapi "github.com/unikorn-cloud/region/pkg/openapi"
+	regionapi "github.com/nscaledev/nscale-sdk-go/region"
 )
 
 type InstanceFlavorModel struct {